CharacterGrid utils class added for Day3 sum calculation
Day3 rewritten

still incorrect. was pretty drained after a long day and decided to come back to it, not sure if i understood what the puzzle was asking for lol, but i'm sure CharacterGrid will come in handy for later puzzles regardless - will start working on structure soon next time i hop on this, i think.